requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pgvector==0.3.6

# Cache / Queue
redis==5.2.1
celery==5.4.0

# Config
pydantic-settings==2.7.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI
openai==1.59.6
langgraph>=0.2.0
twilio>=9.0.0

# File uploads
python-multipart==0.0.20
aiofiles>=23.0.0
pdfplumber>=0.11.0

# Testing
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
anyio[trio]==4.8.0
